refactor(rentals): group services by use case
