+++ apps/server/src/domain/rentals/index.ts
@@ -5,15 +5,15 @@ export * from "./repository/rental.repository";
 export * from "./repository/return-confirmation.repository";
 export * from "./repository/return-slot.repository";
 export * from "./repository/unique-violation";
-export * from "./services/admin-rentals.service";
-export * from "./services/confirm-return.service";
-export * from "./services/environment-impact-job.service";
-export * from "./services/finalize-rental-return";
-export * from "./services/rental-billing-detail.service";
-export * from "./services/rental-billing-preview.service";
-export * from "./services/rental-command.service";
-export * from "./services/rental-stats.service";
-export * from "./services/rental.service";
-export * from "./services/return-slot.service";
-export * from "./services/staff-rental.service";
-export * from "./services/start-rental.service";
+export * from "./services/commands/confirm-return.service";
+export * from "./services/commands/finalize-rental-return";
+export * from "./services/commands/rental-command.service";
+export * from "./services/commands/return-slot.service";
+export * from "./services/commands/start-rental.service";
+export * from "./services/queries/rental-billing-detail.service";
+export * from "./services/queries/rental-billing-preview.service";
+export * from "./services/queries/rental-stats.service";
+export * from "./services/queries/rental.service";
+export * from "./services/roles/admin-rentals.service";
+export * from "./services/roles/staff-rental.service";
+export * from "./services/workers/environment-impact-job.service";
