+++ apps/server/src/domain/rentals/services/commands/confirm-return/confirm-return.types.ts
