+++ apps/server/src/domain/rentals/services/commands/start-rental/start-rental.types.ts
