+++ apps/server/src/domain/rentals/services/queries/rental-stats-time.ts
