+++ apps/server/src/domain/rentals/services/queries/rental-stats.service.ts
@@ -1,14 +1,14 @@
 import { Context, Effect, Layer } from "effect";
 
-import type { RentalRepositoryError } from "../domain-errors";
+import type { RentalRepositoryError } from "../../domain-errors";
 import type {
   RentalDashboardSummary,
   RentalRevenueGroupBy,
   RentalRevenueStats,
   RentalSummaryStats,
-} from "../models";
+} from "../../models";
 
-import { RentalAnalyticsRepository } from "../repository/rental-analytics.repository";
+import { RentalAnalyticsRepository } from "../../repository/rental-analytics.repository";
 import { aggregateRentalStatusCounts } from "./rental-counts";
 import {
   comparePercentage,
