+++ apps/server/src/domain/rentals/services/shared/unique-violation-mapper.ts
@@ -1,8 +1,8 @@
 import { Option } from "effect";
 
-import type { RentalServiceFailure } from "../domain-errors";
+import type { RentalServiceFailure } from "../../domain-errors";
 
-import { ActiveRentalExists, BikeAlreadyRented } from "../domain-errors";
+import { ActiveRentalExists, BikeAlreadyRented } from "../../domain-errors";
 
 function normalizeConstraintList(
   constraint: string[] | string | undefined,
