+++ apps/server/src/domain/rentals/services/workers/environment-impact-job.service.ts
