+++ apps/server/src/http/controllers/rentals/agency.controller.ts
@@ -11,7 +11,7 @@ import {
   staffApproveBikeSwapRequest,
   staffGetChangeBikeDetail,
   staffRejectBikeSwapRequest,
-} from "@/domain/rentals/services/staff-rental.service";
+} from "@/domain/rentals/services/roles/staff-rental.service";
 import { withLoggedCause } from "@/domain/shared";
 import {
   toContractAdminRentalDetail,
